% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{add_transitive_edges}
\alias{add_transitive_edges}
\title{Find inputs to a nest}
\usage{
add_transitive_edges(m, bv, deps, final, parent)
}
\arguments{
\item{m}{The current monadic node}

\item{bv}{A named list of bound variables}

\item{deps}{A mapping local variables to bound variable dependencies}
}
\description{
Find inputs to a nest
}
\keyword{internal}
