% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewriters.R
\name{apply_rewriters}
\alias{apply_rewriters}
\title{Apply rewriters to an Rmonad}
\usage{
apply_rewriters(x, meta = .single_meta(x))
}
\arguments{
\item{x}{The Rmonad}

\item{meta}{A metadata list}
}
\description{
Rewriters are functions stored in an Rmonad's metadata list that operate on
an Rmonad after it has evaluated its code.
}
