% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics.R
\name{as.list.Rmonad}
\alias{as.list.Rmonad}
\title{Convert an rmonad object to a list of monads}
\usage{
\method{as.list}{Rmonad}(x, ...)
}
\arguments{
\item{x}{An Rmonad object}

\item{...}{Additional arguments (unused)}
}
\description{
Convert an rmonad object to a list of monads
}
