% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as.monadR}
\alias{as.monadR}
\title{Lift a value into the report monad if not in one already}
\usage{
as.monadR(x)
}
\arguments{
\item{x}{A value}
}
\value{
Value wrapped in a report monad
}
\description{
Lift a value into the report monad if not in one already
}
