% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_x-to-monad.R
\name{as_monad}
\alias{as_monad}
\title{Evaluate an expression into an Rmonad, capturing events}
\usage{
as_monad(expr, desc = NULL, doc = NULL)
}
\arguments{
\item{expr}{An expression}

\item{desc}{A name to assign to the code slot}

\item{doc}{A docstring to associate with the monad}
}
\value{
Rmonad object
}
\description{
If the value is already an Rmonad, the existing value is returned.
}
\examples{
as_monad(stop(1))
as_monad(1:10)
as_monad(5 \%>>\% sqrt)
}
