% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as_rmonad}
\alias{as_rmonad}
\title{Lift a value into the report monad if not in one already}
\usage{
as_rmonad(x, ...)
}
\arguments{
\item{x}{A value}

\item{...}{extra state information for \code{pass}}
}
\value{
Value wrapped in a report monad
}
\description{
Lift a value into the report monad if not in one already
}
