% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bindMerge}
\alias{bindMerge}
\title{Apply a function to several arguments}
\usage{
bindMerge(xs, f)
}
\arguments{
\item{xs}{A list of inputs to f, these may or may not be in report monads}

\item{f}{some function of xs}
}
\value{
a result in a report monad
}
\description{
This function is a multivariate wrapper for bind. All elements in the list
of input arguments, \code{xs}, are first cast into the \code{monadR} class
(if not in it already). Then these are merged into one report monad. This
report monad is then sent to \code{bind}. You can think of it as merging
many arguments into a single tuple of arguments.
}
