% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{combine}
\alias{combine}
\title{Merge list of reports into one}
\usage{
combine(ms, keep_history = TRUE)
}
\arguments{
\item{ms}{List of report}

\item{keep_history}{Merge the histories of all monads}
}
\value{
A combined report
}
\description{
Merge list of reports into one
}
