% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_helpers.R
\name{const}
\alias{const}
\title{Ignore the first input, return the second}
\usage{
const(x, r)
}
\arguments{
\item{x}{ignored value}

\item{r}{replacing value}
}
\description{
This function can be used to change the value in the lhs of a monadic
sequence
}
\seealso{
Other helper_functions: \code{\link{false_as_error}},
  \code{\link{false}}, \code{\link{null_as_error}},
  \code{\link{toss}}, \code{\link{true}}
}
