% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{crunch}
\alias{crunch}
\title{Cache all large values that are stored in memory}
\usage{
crunch(m)
}
\arguments{
\item{m}{Rmonad object}
}
\description{
Cache all large values that are stored in memory
}
\examples{
poo = list(); document() #####!!!!!!!!!!!!!!!!!!!!!!!!
set.seed(42)
m <- as_monad(runif(1e6), tag="a") \%>>\%
     sqrt \%>\% tag("b") \%>>\%
     log \%>\% tag("c") \%>>\% prod(2) \%>>\% prod(3)
m1 <- crunch(m)
get_value(m,  1:3) \%>\% lapply(head)
get_value(m1, 1:3) \%>\% lapply(head)
}
