% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-monad.R
\name{doc}
\alias{doc}
\title{Add documentation to a monad}
\usage{
doc(m, ...)
}
\arguments{
\item{m}{An Rmonad}

\item{...}{any number of strings}
}
\value{
The input Rmonad with an added doc string
}
\description{
Add documentation to a monad
}
\seealso{
Other monad-to-monad: \code{\link{forget}}
}
