% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_monad-to-x.R
\name{esc}
\alias{esc}
\title{Returns the value of a monad holds}
\usage{
esc(m)
}
\arguments{
\item{m}{An Rmonad}
}
\description{
If the monad is in the passing state, this will bee a pure value. If the
monad is in the failing state, it will raise an appropriate error.
}
\seealso{
Other monad-to-x: \code{\link{missues}},
  \code{\link{mtabulate}}
}
