% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nse.R
\name{extract_docstring}
\alias{extract_docstring}
\title{Extract a docstring from an anonymous function}
\usage{
extract_docstring(expr)
}
\arguments{
\item{expr}{an expression with an optional docstring as the first statement}
}
\value{
a list of two elements, the expression with the docstring removed,
        and the docstring itself
}
\description{
Extract a docstring from an anonymous function
}
\examples{
expr <- substitute(
  {
    "this is the docstring"
    5 * 32
  }
)
extract_docstring(expr)
}
\keyword{internal}
