% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nse.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract docstring and meta data from an anonymous function}
\usage{
extract_metadata(expr)
}
\arguments{
\item{expr}{an expression with an optional docstring as the first statement}
}
\value{
a list of three elements, the expression with the docstring and
metadata removed, the docstring itself, and the metadata.
}
\description{
Currently the docstring must come first and the metadata list must come
second. Both must preceed any other code in the function. These constraints
may be relaxed in the future.
}
\examples{
expr <- substitute(
  {
    "this is the docstring"
    list(foo="this is meta data")
    5 * 32
  }
)
extract_metadata(expr)
}
\keyword{internal}
