% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fail}
\alias{fail}
\title{Load a failure message into the monad}
\usage{
fail(x, s)
}
\arguments{
\item{x}{A value (which is be ignored)}

\item{s}{An error message}
}
\value{
A failing monad report
}
\description{
Load a failure message into the monad
}
