% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{fail_cache}
\alias{fail_cache}
\title{Represent a dummy value for a node downstream of a failing node}
\usage{
fail_cache()
}
\value{
A function that represents an unrun node
}
\description{
Returns a ValueManager that represents a dummy value for a node downstream
of a failing node. Unlike \code{void_cache}, this presence of this manager
in a pipeline is not pathological, so does not raise a warning by default.
}
\seealso{
Other cache: \code{\link{clear_cache}},
  \code{\link{make_cacher}}, \code{\link{make_recacher}},
  \code{\link{memory_cache}}, \code{\link{no_cache}},
  \code{\link{void_cache}}
}
