% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{false}
\alias{false}
\title{Return false for all input}
\usage{
false(...)
}
\arguments{
\item{...}{whatever}
}
\description{
Return false for all input
}
\seealso{
Other helper_functions: \code{\link{const}},
  \code{\link{false_as_error}},
  \code{\link{null_as_error}}, \code{\link{toss}},
  \code{\link{true}}
}
