% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helpers.R
\name{false_as_error}
\alias{false_as_error}
\title{Make NULL values an error}
\usage{
false_as_error(x)
}
\arguments{
\item{x}{Input value}
}
\description{
Make NULL values an error
}
\seealso{
Other helper_functions: \code{\link{const}},
  \code{\link{false}}, \code{\link{ident}},
  \code{\link{null_as_error}}, \code{\link{toss}},
  \code{\link{true}}
}
