% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helpers.R
\name{first}
\alias{first}
\title{Given two arguments, return the first}
\usage{
first(x, y)
}
\arguments{
\item{x}{anything}

\item{y}{anything}
}
\description{
Given two arguments, return the first
}
\seealso{
Other help_functions: \code{\link{second}}
}
