% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fmap}
\alias{fmap}
\title{This function is appropriate when function f will not fail}
\usage{
fmap(m, f)
}
\arguments{
\item{m}{Either a monadic error container or a simple value}

\item{f}{A function to apply to the value contained by the container}
}
\description{
This function is appropriate when function f will not fail
}
