% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_monad-to-monad.R
\name{forget}
\alias{forget}
\title{Wipe a monad's history}
\usage{
forget(m)
}
\arguments{
\item{m}{An Rmonad}
}
\value{
The input Rmonad with history erased
}
\description{
Wipe a monad's history
}
\seealso{
Other monad-to-monad: \code{\link{doc}},
  \code{\link{store}}
}
