% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_magic.R
\name{get_dependency_matrix}
\alias{get_dependency_matrix}
\title{Get dependencies of local variables on inputs}
\usage{
get_dependency_matrix(declarations, bound_vars)
}
\arguments{
\item{declarations}{A list of declarations}

\item{bound_vars}{Character vector of variables names that are bound as
arguments to the function}
}
\value{
logical matrix
}
\description{
Get dependencies of local variables on inputs
}
