% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gff}
\alias{gff}
\title{Data for GFF processing vignette}
\format{List}
\usage{
gff
}
\description{
Contains 4 files:
\enumerate{
  \item good - a valid GFF string
  \item not_a_gff1 - a string that isn't a GFF file at all
  \item invalid_type - a table with invalid types
  \item good_result - the final pipeline produced using the good gff
}
}
\keyword{datasets}
