% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gff_str}
\alias{gff_str}
\title{Data for GFF processing vignette}
\format{List}
\usage{
gff_str
}
\description{
Contains 4 files:
\enumerate{
    \item synmap - hand made synteny map
    \item qgff   - set of intervals relative to the query
    \item tgff   - set of intervals relative to the target
    \item hitmap - map to filter through the main synteny map
    \item qgen   - lengths of query contigs/scaffolds/chromosomes
    \item tgen   - lengths of target contigs/scaffolds/chromosomes
}
}
\keyword{datasets}
