% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{\%>>=\%}
\alias{\%>>=\%}
\title{Infix version of \code{bind}}
\usage{
l \%>>=\% r
}
\arguments{
\item{l}{left-hand side}

\item{r}{right-hand side}
}
\value{
result of bind(l, r)
}
\description{
Infix version of \code{bind}
}
