% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{id}
\alias{id}
\title{The identity function}
\usage{
id(x)
}
\arguments{
\item{x}{anything}
}
\description{
The identity function
}
\examples{
# pull a value out of failure
1:10 \%$>\% colSums \%?>\% id
# here it just does nothing
cars \%$>\% colSums \%?>\% id
}
