% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_accessors.R
\name{is_rmonad}
\alias{is_rmonad}
\title{Determine wether something is an Rmonad object}
\usage{
is_rmonad(m)
}
\arguments{
\item{m}{Rmonad object}
}
\value{
logical TRUE if m is an Rmonad
}
\description{
Determine wether something is an Rmonad object
}
