% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monadic_eval.R
\name{lsmeval}
\alias{lsmeval}
\title{Safely builds a list of monads from a list of data}
\usage{
lsmeval(xs, ...)
}
\arguments{
\item{xs}{An unevaluated list of data}

\item{...}{Additional arguments for combine}
}
\value{
A list of Rmonads
}
\description{
Safely builds a list of monads from a list of data
}
\examples{
lsmeval( list(1:10, stop(1)) )
}
