% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_accessors.R
\name{m_delete_value}
\alias{m_delete_value}
\title{Delete a node's value}
\usage{
m_delete_value(m)
}
\arguments{
\item{m}{Rmonad object}
}
\description{
This is not the same as setting the value to NULL. Deleting tells
\code{rmonad} that no value is cached. If the value is accessed later, a
warning is raised. In contrast, setting the value to NULL will result in
\code{rmonad} thinking that the result of the computation was NULL.
}
