% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{make_cacher}
\alias{make_cacher}
\title{Make Cacher object}
\usage{
make_cacher(f_path = function() getOption("rmonad.cache_dir"),
  f_save = saveRDS, f_get = readRDS, f_del = unlink,
  f_ext = function(cls) ".Rdata")
}
\arguments{
\item{f_path}{A function for finding the directory in which to cache results}

\item{f_save}{function of x and filename that saves x to the path filename}

\item{f_get}{function of filename that retrieves the cached data}

\item{f_del}{function of filename that deletes the cached data}

\item{f_ext}{function of class(x) that determines the filename extension}
}
\value{
A function that builds a local cache function for a value
}
\description{
Make Cacher object
}
\seealso{
Other cache: \code{\link{clear_cache}},
  \code{\link{fail_cache}}, \code{\link{make_recacher}},
  \code{\link{memory_cache}}, \code{\link{no_cache}},
  \code{\link{void_cache}}
}
\concept{cache}
