% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{make_local_cacher}
\alias{make_local_cacher}
\title{Make a function of x that caches data locally}
\usage{
make_local_cacher(path)
}
\arguments{
\item{path}{A directory in which to cache results}
}
\value{
A function that builds a local cache function for a value
}
\description{
Make a function of x that caches data locally
}
\examples{
\dontrun{
  foo <- 45
  cacher <- make_local_cacher(".")
  foo_ <- cacher(45)
  rm(foo)
  foo_@get()
}
}
