% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{makeCacher}
\alias{makeCacher}
\title{Make a function that taks an Rmonad and recaches it}
\usage{
makeCacher(cacher)
}
\arguments{
\item{cacher}{A function of a data value}
}
\description{
Make a function that taks an Rmonad and recaches it
}
