% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{makeRecacher}
\alias{makeRecacher}
\title{Make a function that taks an Rmonad and recaches it}
\usage{
makeRecacher(cacher)
}
\arguments{
\item{cacher}{A function of a data value}
}
\value{
A function that swaps the cache function of an Rmonad
}
\description{
Make a function that taks an Rmonad and recaches it
}
\examples{
\dontrun{
  recacher <- makeRecacher(makeLocalCacher("."))
  m <- iris \%>>\% summary \%>\% recacher
  # load the data from a local file
  m_value(m)

  recacher <- makeRecacher(memoryCache)
  m <- iris \%>>\% summary \%>\% recacher
  # load the data from memory
  m_value(m)
}
}
