% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mcombine}
\alias{mcombine}
\title{Combine two reports into one}
\usage{
mcombine(m1, m2)
}
\arguments{
\item{m1}{Report monad}

\item{m2}{Report monad}
}
\value{
A report monad with combines m1 and m2
}
\description{
This is the associative binary operator that, together with the zero element
\code{mzero}, define the report monad as a monoid.
}
\details{
Since each slot in the monadR class is a list, two reports can be
combined simple by appending each list in m2 to the corresponding list in
m1.
}
