% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{memory_cache}
\alias{memory_cache}
\title{Store a value in memory}
\usage{
memory_cache(x)
}
\arguments{
\item{x}{Value to be stored}
}
\value{
A function that returns a value stored in memory
}
\description{
Store a value in memory
}
\examples{
foo <- 45
foo_proxy <- memory_cache(foo)
foo
foo_proxy@get()
}
