% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{memoryCache}
\alias{memoryCache}
\title{Store a value in memory}
\usage{
memoryCache(x)
}
\arguments{
\item{x}{Value to be stored}
}
\value{
A function that returns a value stored in memory
}
\description{
Store a value in memory
}
\examples{
foo <- 45
foo_proxy <- memoryCache(foo)
foo
foo_proxy@get()
}
