% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_monad-to-monad.R
\name{mfind}
\alias{mfind}
\title{Extract a monad by ID}
\usage{
mfind(m, id, isolate = FALSE)
}
\arguments{
\item{m}{An Rmonad}

\item{id}{a single integer ID}

\item{isolate}{logical, should the context of the returned monad be removed}
}
\value{
a single Rmonad
}
\description{
If the id is not found, an error is raised
}
\seealso{
Other monad-to-monad: \code{\link{doc}},
  \code{\link{forget}}, \code{\link{store}}
}
