% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{class}
\name{monadR-class}
\alias{monadR-class}
\alias{monadR}
\title{The eponymous monad type}
\description{
The eponymous monad type
}
\section{Slots}{

\describe{
\item{\code{x}}{List of values produced if the parent computation succeeded}

\item{\code{errors}}{List of errors accumulated so far}

\item{\code{warnings}}{List of warnings accumulated so far}

\item{\code{notes}}{List of notes accumulated so far}

\item{\code{OK}}{TRUE if the report is currently passing}
}}

