% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{mreport}
\alias{mreport}
\title{Convert a pipeline to Rmarkdown}
\usage{
mreport(m, prefix = "report")
}
\arguments{
\item{m}{An Rmonad}

\item{prefix}{A file prefix for the generated report}
}
\description{
Convert a pipeline to Rmarkdown
}
\examples{
\dontrun{
mreport(-1:2 \%>>\% log \%>>\% sqrt \%__\% "asdf" \%>>\% sqrt)
}
}
\seealso{
Other from_Rmonad: \code{\link{esc}},
  \code{\link{missues}}, \code{\link{mtabulate}}
}
