% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monadic_eval.R
\name{mrun}
\alias{mrun}
\title{Evaluate an expression into an Rmonad, capturing events}
\usage{
mrun(expr, desc = NULL)
}
\arguments{
\item{expr}{An expression}

\item{desc}{A name to assign to the code slot}
}
\value{
Rmonad object
}
\description{
If the value is already an Rmonad, the existing value is returned.
}
\examples{
mrun(stop(1))
mrun(1:10)
mrun(5 \%>>\% sqrt)
}
