% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{mtabulate}
\alias{mtabulate}
\title{Make tabular summary of a monadic chain}
\usage{
mtabulate(m, recurse_nests = TRUE, code = FALSE)
}
\arguments{
\item{m}{An Rmonad}

\item{recurse_nests}{logical Should the resulting table descend into nested pipelines?}

\item{code}{logical Should the code by included?}
}
\description{
Make tabular summary of a monadic chain
}
\seealso{
Other monad-to-x: \code{\link{as_dgr_graph}},
  \code{\link{esc}}, \code{\link{missues}},
  \code{\link{mreport}}
}
