% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mzero}
\alias{mzero}
\title{The monoid zero element for monad report}
\usage{
mzero()
}
\description{
The report monad is also a monoid. Which means there is 1) an associative
operator that can merge reports (\code{mcombine}) and 2) a zero element.
\code{mzero} represents the zero element.
}
