% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nothing}
\alias{nothing}
\title{Do nothing}
\usage{
nothing(...)
}
\arguments{
\item{...}{anything}
}
\value{
nothing
}
\description{
Do nothing
}
\seealso{
Other help_functions: \code{\link{first}},
  \code{\link{second}}
}
