% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_helpers.R
\name{null_as_error}
\alias{null_as_error}
\title{Make NULL values an error}
\usage{
null_as_error(x)
}
\arguments{
\item{x}{Input value}
}
\description{
Currently not exported.
}
\seealso{
Other helper_functions: \code{\link{const}},
  \code{\link{false_as_error}}, \code{\link{false}},
  \code{\link{toss}}, \code{\link{true}}
}
