% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pass}
\alias{pass}
\title{Load a value into the report monad}
\usage{
pass(x, desc = NULL)
}
\arguments{
\item{x}{The result of a successful computation}

\item{desc}{An optional description of the source}
}
\value{
The result wrapped in the report monad
}
\description{
Load a value into the report monad
}
