% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics.R
\name{print.Rmonad}
\alias{print.Rmonad}
\title{Rmonad print generic function}
\usage{
\method{print}{Rmonad}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{An Rmonad object}

\item{verbose}{logical print verbose output (include benchmarking)}

\item{...}{Additional arguments (unused)}
}
\description{
Rmonad print generic function
}
