% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{record-class}
\alias{record-class}
\alias{record}
\title{A record of past events}
\description{
A record of past events
}
\section{Slots}{

\describe{
\item{\code{x}}{A list of 0 or 1 elements, containing either nothing or data}

\item{\code{code}}{A string showing the function that created record's report}

\item{\code{errors}}{List of errors accumulated so far}

\item{\code{warnings}}{List of warnings accumulated so far}

\item{\code{notes}}{List of notes accumulated so far}
}}

