% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{relink_node}
\alias{relink_node}
\title{The recursive function for adding dependencies}
\usage{
relink_node(m, bv, deps)
}
\arguments{
\item{m}{The current monadic node}

\item{bv}{A named list of bound variables}

\item{deps}{A mapping local variables to bound variable dependencies}
}
\description{
The recursive function for adding dependencies
}
\keyword{internal}
