% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{Rmonad-class}
\alias{Rmonad-class}
\alias{Rmonad}
\title{The eponymous monad type}
\description{
The eponymous monad type
}
\section{Slots}{

\describe{
\item{\code{x}}{List of values produced if the parent computation succeeded}

\item{\code{stage}}{The active record}

\item{\code{history}}{A list of past records}

\item{\code{OK}}{TRUE if the report is currently passing}
}}

