% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rmonad_accessors}
\alias{rmonad_accessors}
\alias{m_history}
\alias{m_OK}
\alias{m_value}
\alias{m_code}
\alias{m_errors}
\alias{m_warnings}
\alias{m_notes}
\alias{m_doc}
\alias{m_branch}
\alias{m_history<-}
\alias{m_OK<-}
\alias{m_value<-}
\alias{m_code<-}
\alias{m_errors<-}
\alias{m_warnings<-}
\alias{m_notes<-}
\alias{m_doc<-}
\alias{m_branch<-}
\alias{app_history}
\alias{app_warnings}
\alias{app_notes}
\alias{app_branch}
\title{Get, set, and append Rmonad fields}
\usage{
m_history(m)

m_OK(m)

m_value(m)

m_code(m)

m_errors(m)

m_warnings(m)

m_notes(m)

m_doc(m)

m_branch(m)

m_history(m) <- value

m_OK(m) <- value

m_value(m) <- value

m_code(m) <- value

m_errors(m) <- value

m_warnings(m) <- value

m_notes(m) <- value

m_doc(m) <- value

m_branch(m) <- value

app_history(m, value)

app_warnings(m, value)

app_notes(m, value)

app_branch(m, value)
}
\arguments{
\item{m}{The Rmonad}

\item{value}{Value to replace or append current value}
}
\description{
Get, set, and append Rmonad fields
}
