% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{rmonad_getters}
\alias{rmonad_getters}
\alias{get_parents}
\alias{get_dependents}
\alias{get_nest}
\alias{get_prior}
\alias{get_nest_depth}
\alias{get_value}
\alias{get_id}
\alias{get_OK}
\alias{get_code}
\alias{get_error}
\alias{get_warnings}
\alias{get_notes}
\alias{get_doc}
\alias{get_meta}
\alias{get_time}
\alias{get_mem}
\alias{get_summary}
\title{Vectorized getters for public Rmonad fields}
\usage{
get_parents(m, index = .get_ids(m))

get_dependents(m, index = .get_ids(m))

get_nest(m, index = .get_ids(m))

get_prior(m, index = .get_ids(m))

get_nest_depth(m, index = .get_ids(m))

get_value(m, warn = TRUE, index = .get_ids(m))

get_id(m, index = .get_ids(m))

get_OK(m, index = .get_ids(m))

get_code(m, index = .get_ids(m))

get_error(m, index = .get_ids(m))

get_warnings(m, index = .get_ids(m))

get_notes(m, index = .get_ids(m))

get_doc(m, index = .get_ids(m))

get_meta(m, index = .get_ids(m))

get_time(m, index = .get_ids(m))

get_mem(m, index = .get_ids(m))

get_summary(m, index = .get_ids(m))
}
\arguments{
\item{m}{An Rmonad object}

\item{index}{Selection of indices to extract (all by default). The indices
may be a vector of integers, node names, or igraph vertices
(\code{igraph.vs}).}

\item{warn}{logical In get_value, raise a warning on an attempt to access an uncached node}

\item{...}{Additional arguments}
}
\description{
Vectorized getters for public Rmonad fields
}
