% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{rmonad_meta}
\alias{rmonad_meta}
\alias{store}
\alias{esc}
\alias{forget}
\alias{unbranch}
\alias{doc}
\alias{unstore}
\alias{uncode}
\title{Functions on monads}
\usage{
store(m)

esc(m)

forget(m)

unbranch(m)

doc(m, ...)

unstore(m)

uncode(m)
}
\arguments{
\item{m}{An Rmonad}

\item{...}{additional arguments}
}
\description{
These functions act on Rmonad objects, rather than specifically the values
they contain.
}
\details{
\code{store} stores the monads value in the monads state, so that it is
passed down the chain.

\code{esc} returns the value of a monad or, if the monad is in the failing
state, raises on appropriate error.

\code{forget} wipes a monads history.

\code{unbranch} returns a list of monads, one for each branch

\code{doc} adds a documentation section to the monad.

\code{unstore} retrieves all stored values in history (does not recurse into
branches). Maybe it should.
}
