% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{rmonad_printers}
\alias{rmonad_printers}
\alias{print.record}
\alias{print.rmonad}
\title{Special rmonad_printerss}
\usage{
\method{print}{record}(x, ...)

\method{print}{rmonad}(x, ...)
}
\arguments{
\item{x}{An object to print}

\item{...}{Additional arguments going to God knows where}
}
\description{
Special rmonad_printerss
}
