% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrappers.R
\name{rmonad_unwrap}
\alias{rmonad_unwrap}
\alias{unstore}
\alias{uncode}
\alias{unerror}
\alias{unwarnings}
\alias{unnotes}
\alias{undoc}
\alias{unbranch}
\title{Functions for extracting information from Rmonad history}
\usage{
unstore(m)

uncode(m)

unerror(m)

unwarnings(m)

unnotes(m)

undoc(m)

unbranch(m)
}
\arguments{
\item{m}{An Rmonad}
}
\description{
Each of these functions returns a list of the values for the eponymous slot
(e.g. uncode returns the code run for each past stage).
}
\details{
\code{unbranch} is recursive, returning a monad for each branch. The other
functions are currently not recursive (do not look into branches). I will
change this eventualy.
}
