% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_helpers.R
\name{second}
\alias{second}
\title{Given two arguments, return the second}
\usage{
second(x, y)
}
\arguments{
\item{x}{anything}

\item{y}{anything}
}
\description{
Given two arguments, return the second
}
\seealso{
Other help_functions: \code{\link{first}}
}
