% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{splice_function}
\alias{splice_function}
\title{Take a monadic bind operation's result and splice histories}
\usage{
splice_function(f, m, ms, ...)
}
\arguments{
\item{f}{The function}

\item{m}{The monadic result of running f(ms)}

\item{ms}{The list of inputs passed to f}

\item{...}{additional arguments passed to add_transitive_edges}
}
\description{
We need to link input variables to the nodes in the nested pipeline that use
them.
}
