% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_monad-to-monad.R
\name{store}
\alias{store}
\title{Store a monad's value}
\usage{
store(m)
}
\arguments{
\item{m}{An Rmonad}
}
\value{
The input Rmonad with its value tagged for storage
}
\description{
Store a monad's value
}
\seealso{
Other monad-to-monad: \code{\link{doc}},
  \code{\link{extract_by_id}}, \code{\link{forget}}
}
\keyword{internal}
