% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{tag}
\alias{tag}
\title{Set the tag of an Rmonad object}
\usage{
tag(m, value, index = m@head)
}
\arguments{
\item{m}{Rmonad object}

\item{value}{string specifying a new tag for the given nodes}

\item{index}{character or integer vector, specifying the nodes which will be
assigned the new tag}
}
\value{
Rmonad object with new tags
}
\description{
Set the tag of an Rmonad object
}
