% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{toss}
\alias{toss}
\title{Take input and do nothing with it}
\usage{
toss(...)
}
\arguments{
\item{...}{whatever}
}
\description{
Take input and do nothing with it
}
\seealso{
Other helper_functions: \code{\link{const}},
  \code{\link{false_as_error}}, \code{\link{false}},
  \code{\link{null_as_error}}, \code{\link{true}}
}
