% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_helpers.R
\name{true}
\alias{true}
\title{Return true for all input}
\usage{
true(...)
}
\arguments{
\item{...}{whatever}
}
\description{
Return true for all input
}
\seealso{
Other helper_functions: \code{\link{const}},
  \code{\link{false_as_error}}, \code{\link{false}},
  \code{\link{ident}}, \code{\link{null_as_error}},
  \code{\link{toss}}
}
