% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{unbranch}
\alias{unbranch}
\title{Return each independent branch of the pipeline}
\usage{
unbranch(m)
}
\arguments{
\item{m}{An Rmonad}
}
\value{
A list of Rmonads, one for each branch
}
\description{
Branches are generated by the \code{\%>^\%} operator. The allow a value to
be sent into a new branch of the pipeline, which can fail independently,
without propagating to other pieces of the program.
}
