% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-monad.R
\name{unnest}
\alias{unnest}
\title{Remove a level of nesting in an Rmonad}
\usage{
unnest(m)
}
\arguments{
\item{m}{An Rmonad}
}
\description{
Remove a level of nesting in an Rmonad
}
\seealso{
Other monad-to-monad: \code{\link{doc}}
}
