% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{view}
\alias{view}
\title{Set the head of an Rmonad to a particular tag}
\usage{
view(m, ...)
}
\arguments{
\item{m}{Rmonad object}

\item{...}{one or more tag strings specifying a unique node in the pipeline}
}
\value{
Rmonad object with head reset
}
\description{
Will split on '/'
}
\examples{
library(magrittr)
m <- 256 \%v>\% sqrt \%>\% tag('a', 'b') \%v>\% sqrt
esc(view(m, 'a/b'))
funnel(view(m, 'a'), m) \%*>\% sum
}
