% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{viewIDs}
\alias{viewIDs}
\title{Return a list of Rmonad objects at these positions}
\usage{
viewIDs(m, ids)
}
\arguments{
\item{m}{rmonad object}

\item{ids}{integer vector index}
}
\description{
Return a list of Rmonad objects at these positions
}
