% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{views}
\alias{views}
\title{Get a list of Rmonad objects matching the given tag}
\usage{
views(m, ...)
}
\arguments{
\item{m}{Rmonad object}

\item{...}{one or more tags}
}
\value{
list of Rmonad objects
}
\description{
Get a list of Rmonad objects matching the given tag
}
\examples{
library(magrittr)
1 \%>>\% prod(2) \%>\% tag('a/b') \%>>\%
       prod(2) \%>\% tag('a/c') \%>>\%
       prod(2) \%>\% tag('a/c') \%>>\%
       prod(2) \%>\% tag('g/a') -> m
views(m, 'a')
}
