% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{void_cache}
\alias{void_cache}
\title{Represent a value that has not been set}
\usage{
void_cache()
}
\value{
A function that represents a void, uncached value
}
\description{
This is the default value of RmonadData@value. It should always be replaced
shortly after the object is created, thus should only be encountered if 1)
the user is directly creating RmonadData objects (in which case they should
be spoken to sternly) or 2) there is a bug in rmonad.
}
