% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{voidCache}
\alias{voidCache}
\title{Represent a value that has not been set}
\usage{
voidCache()
}
\description{
This is the default value of RmonadNode@value. It should always be replaced
shortly after the object is created, thus should only be encountered if 1)
the user is directly creating RmonadNode objects (in which case they should
be spoken to sternly) or 2) there is a bug in rmonad.
}
