% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{warn}
\alias{warn}
\title{Append a warning message onto the monad}
\usage{
warn(m, s)
}
\arguments{
\item{m}{A report monad}

\item{s}{A string describing a warning}
}
\value{
A report monad with a new warning appended
}
\description{
Append a warning message onto the monad
}
