% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monadic_eval.R
\name{as_monad}
\alias{as_monad}
\title{Evaluate an expression into an Rmonad, capturing events}
\usage{
as_monad(expr, desc = NULL)
}
\arguments{
\item{expr}{An expression}

\item{desc}{A name to assign to the code slot}
}
\value{
Rmonad object
}
\description{
If the value is already an Rmonad, the existing value is returned.
}
\examples{
as_monad(stop(1))
as_monad(1:10)
as_monad(5 \%>>\% sqrt)
}
