% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bind}
\alias{bind}
\title{Apply f to the contents of a monad and merge messages}
\usage{
bind(x, f, print_in = FALSE, record_in = FALSE, branch = FALSE,
  discard = FALSE)
}
\arguments{
\item{x}{The input, may or may not be a monad report}

\item{f}{A function of the value contained in x}

\item{print_in}{logical - should the input be printed?}

\item{record_in}{logical - should the input be recorded?}

\item{branch}{Store the output in the branch slot}

\item{discard}{Ignore the result of the rhs function, passon the left}
}
\value{
A monad report
}
\description{
This function uses non-standard evaluation to insert x into f as the first
positional argument. This allows specialization of f, but also prevents
higher-order voodoo from being performed.
}
