% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bind}
\alias{bind}
\title{Apply f to the contents of a monad and merge messages}
\usage{
bind(x, f, on_entry = ident, bind_if = function(m) {     m_OK(m) },
  record_in = FALSE, branch = FALSE, discard_out = FALSE,
  discard_in = FALSE, handle = FALSE)
}
\arguments{
\item{x}{The input, may or may not be a monad report}

\item{f}{A function of the value contained in x}

\item{on_entry}{f(m,x) an action to perform on entry, returns m}

\item{bind_if}{f(m) bind rhs to lhs if TRUE}

\item{record_in}{logical - should the input be recorded?}

\item{branch}{Store the output in the branch slot}

\item{discard_out}{Ignore the result of the rhs function, passon the left}

\item{discard_in}{Ignore the input}

\item{handle}{Operate on a failed state}
}
\value{
A monad report
}
\description{
This function should not usually be used directly. Rather you should use the
infix operators. They all wrap this function.
}
\details{
This function uses non-standard evaluation to insert x into f as the first
positional argument. This allows specialization of f, but also prevents
higher-order voodoo from being performed.
}
