% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bind}
\alias{bind}
\title{Apply f to the contents of a monad and merge messages}
\usage{
bind(x, f)
}
\arguments{
\item{x}{The input, may or may not be a monad report}

\item{f}{A function of the value contained in x}
}
\value{
A monad report
}
\description{
Apply f to the contents of a monad and merge messages
}
