% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bind}
\alias{bind}
\title{Apply f to the contents of a monad and merge messages}
\usage{
bind(x, f)
}
\arguments{
\item{x}{The input, may or may not be a monad report}

\item{f}{A function of the value contained in x}
}
\value{
A monad report
}
\description{
This function uses non-standard evaluation to insert x into f as the first
positional argument. This allows specialization of f, but also prevents
higher-order voodoo from being performed.
}
\examples{
bind(5, runif(min=10, max=20))
}
