% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_bind.R
\name{bind}
\alias{bind}
\title{Apply f to the contents of a monad and merge messages}
\usage{
bind(x, f, entry_lhs_transform = entry_lhs_transform_default,
  bind_if = function(m) m_OK(m), bind_else = function(...) {     NULL },
  emit = emit_default, m_on_bind = function(x, ...) {     x },
  io_combine = default_combine, bind_args = function(m) list(m_value(m, warn
  = FALSE)), bind_monad = function(m) list(m), expect_rhs_function = TRUE,
  envir = parent.frame())
}
\arguments{
\item{x}{The input, may or may not be a monad report}

\item{f}{A function of the value contained in x}

\item{entry_lhs_transform}{f(m,x,...) a transform of the lhs called on entry}

\item{bind_if}{f(m) bind rhs to lhs if TRUE}

\item{bind_else}{f(m,f) action to take if bind_if is FALSE}

\item{emit}{f(i,o) Emit the input or the output}

\item{m_on_bind}{f(m) Action to perform on input monad when binding}

\item{io_combine}{f(m,o) weave m and f(m) into final output}

\item{bind_args}{function to retrieve the arguments}
}
\value{
A monad report
}
\description{
This function should not be used directly. Rather you should use the infix
operators. They all wrap this function.
}
\keyword{internal}
