% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cached values and delete temporary files}
\usage{
clear_cache(m, index = .get_ids(m))
}
\arguments{
\item{m}{Rmonad object}

\item{index}{indices to clear (all indices by default)}
}
\value{
Rmonad object
}
\description{
Clear cached values and delete temporary files
}
\examples{
256 \%v>\% sqrt \%>>\% sqrt \%>>\% sqrt  -> m
m
clear_cache(m)
}
\seealso{
Other cache: \code{\link{fail_cache}},
  \code{\link{make_cacher}}, \code{\link{make_recacher}},
  \code{\link{memory_cache}}, \code{\link{no_cache}},
  \code{\link{void_cache}}
}
