% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_monad-to-monad.R
\name{forget}
\alias{forget}
\title{Clear an rmonad's parents}
\usage{
forget(m)
}
\arguments{
\item{m}{An Rmonad}
}
\value{
The input Rmonad with all parents erased
}
\description{
Clear an rmonad's parents
}
\seealso{
Other monad-to-monad: \code{\link{doc}},
  \code{\link{extract_by_id}}, \code{\link{store}}
}
