% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_helpers.R
\name{ident}
\alias{ident}
\title{The identity function}
\usage{
ident(x, ...)
}
\arguments{
\item{x}{anything}

\item{...}{parameters that will be ignored}
}
\description{
The identity function
}
\examples{
# pull a value out of failure
1:10 \%$>\% colSums \%|>\% ident
# here it just does nothing
cars \%$>\% colSums \%|>\% ident
}
\seealso{
Other helper_functions: \code{\link{const}},
  \code{\link{false_as_error}}, \code{\link{false}},
  \code{\link{null_as_error}}, \code{\link{toss}},
  \code{\link{true}}
}
