% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helpers.R
\name{ident}
\alias{ident}
\title{The identity function}
\usage{
ident(x, ...)
}
\arguments{
\item{x}{anything}

\item{...}{parameters that will be ignored}
}
\description{
The identity function
}
\examples{
# pull a value out of failure
1:10 \%$>\% colSums \%|>\% ident
# here it just does nothing
cars \%$>\% colSums \%|>\% ident
}
