% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_operators.R
\name{infix}
\alias{infix}
\alias{\%>>\%}
\alias{\%v>\%}
\alias{\%*>\%}
\alias{\%>_\%}
\alias{\%>^\%}
\alias{\%^>\%}
\alias{\%|>\%}
\alias{\%||\%}
\alias{\%__\%}
\alias{\%v__\%}
\title{Infix operators}
\usage{
lhs \%>>\% rhs

lhs \%v>\% rhs

lhs \%*>\% rhs

lhs \%>_\% rhs

lhs \%>^\% rhs

lhs \%^>\% rhs

lhs \%|>\% rhs

lhs \%||\% rhs

lhs \%__\% rhs

lhs \%v__\% rhs
}
\arguments{
\item{lhs}{left hand value}

\item{rhs}{right hand value}
}
\description{
Infix monadic sequence operators
}
\details{
See the main package help page (\code{?rmonad}) or the \code{intro} and
\code{cheatsheet} vignettes for more information.
}
