% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{infix}
\alias{infix}
\alias{\%>>\%}
\alias{\%$>\%}
\alias{\%v>\%}
\alias{\%^>\%}
\alias{\%>v\%}
\title{Infix operators}
\usage{
lhs \%>>\% rhs

lhs \%$>\% rhs

lhs \%v>\% rhs

lhs \%^>\% rhs

lhs \%>v\% rhs
}
\arguments{
\item{lhs}{left hand value}

\item{rhs}{right hand value}
}
\description{
Infix monadic sequence operators
}
\details{
\strong{Monadic bind operators}

\tabular{ll}{

\%>>\%    monadic bind

\%v>\%    monadic bind and print input

\%^>\%    monadic bind and record input in monad

\%>v\%    run right-hand for its effects -- bind message, pass value

\%>^\%    like \%>v\% but stores the result in the monad

\%?>\%    if input is error, run rhs on last passing result

\%__\%    ignore the input, replacing the value with the rhs

\%=_\%    assign input to a variable (lfs) and clear

\%=>\%    assign input to a variable (lfs) and pass

}

\strong{Monadic list operators}

\tabular{ll}{

\%,,\%    merges rhs and lhs monads

\%,?\%    if last list element failed, use this one instead

\%,=\%    assign a name the last element in the list

\%,>\%    run rhs function with lhs list of arguments

\%,@\%    curry lhs list of paramters into rhs function

}

\strong{Utility operators}

\tabular{ll}{

\%.\%     compose two functions (math order)

\%|\%     compose two functions (pipeline order)

\%>\%     pass lhs into rhs (from magrittr)

}
}
