% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{is_rmonad}
\alias{is_rmonad}
\title{Determine whether something is an Rmonad object}
\usage{
is_rmonad(m)
}
\arguments{
\item{m}{Rmonad object}
}
\value{
logical TRUE if m is an Rmonad
}
\description{
Determine whether something is an Rmonad object
}
