% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_x-to-monad.R
\name{lsmeval}
\alias{lsmeval}
\title{Safely builds a list of monads from an argument list of expressions}
\usage{
lsmeval(..., keep_history = TRUE)
}
\arguments{
\item{...}{expressions to be wrapped into monads}

\item{keep_history}{Merge the histories of all monads}
}
\value{
A list of Rmonads
}
\description{
Safely builds a list of monads from an argument list of expressions
}
\examples{
lsmeval( 1:10, stop(1), sqrt(-3:3) )
}
