% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{m_delete_value}
\alias{m_delete_value}
\title{Delete a node's value}
\usage{
m_delete_value(m, ...)
}
\arguments{
\item{m}{Rmonad object}

\item{index}{Delete the value contained by this vertex (if NULL, delete head value)}
}
\description{
Delete a node's value
}
