% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{make_local_cacher}
\alias{make_local_cacher}
\title{Make a function of x that caches data locally}
\usage{
make_local_cacher(path = getOption("rmonad.cache_dir"), save = saveRDS,
  get = readRDS, del = unlink, ext = function(cls) ".Rdata")
}
\arguments{
\item{path}{A directory in which to cache results.}

\item{save}{function of x and filename that saves x to the path filename}

\item{get}{function of filename that retrieves the cached data}

\item{del}{function of filename that deletes the cached data}

\item{ext}{function of class(x) that determines the filename extension}

\item{chk}{function of filename that checks existence of the cached data}
}
\value{
A function that builds a local cache function for a value
}
\description{
Make a function of x that caches data locally
}
\examples{
\dontrun{
  foo <- 45
  cacher <- make_local_cacher()
  foo_ <- cacher(45)
  rm(foo)
  foo_@get()
}
}
\seealso{
Other cache: \code{\link{clear_cache}},
  \code{\link{fail_cache}}, \code{\link{make_recacher}},
  \code{\link{memory_cache}}, \code{\link{no_cache}},
  \code{\link{void_cache}}
}
