% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{makeLocalCacher}
\alias{makeLocalCacher}
\title{Make a function of x that caches data locally}
\usage{
makeLocalCacher(path)
}
\arguments{
\item{path}{A directory in which to cache results}
}
\description{
Make a function of x that caches data locally
}
\examples{
\dontrun{
  foo <- 45
  localCacher <- makeLocalCacher(".")
  foo_ <- localCacher(45)
  rm(foo)
  foo_@get()
}
}
