% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{missues}
\alias{missues}
\title{Tabulates all errors, warnings and notes}
\usage{
missues(m, recurse_nests = TRUE)
}
\arguments{
\item{m}{An Rmonad}

\item{recurse_nests}{logical Should the resulting table descend into nested pipelines?}
}
\description{
Tabulates all errors, warnings and notes
}
\seealso{
Other from_Rmonad: \code{\link{esc}},
  \code{\link{mreport}}, \code{\link{mtabulate}}
}
