% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{missues}
\alias{missues}
\title{Tabulates all errors, warnings and notes}
\usage{
missues(m)
}
\arguments{
\item{m}{An Rmonad}
}
\description{
Tabulates all errors, warnings and notes
}
\seealso{
Other monad-to-x: \code{\link{as_dgr_graph}},
  \code{\link{esc}}, \code{\link{mreport}},
  \code{\link{mtabulate}}
}
