% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-monad.R
\name{missues}
\alias{missues}
\title{Tabulates all errors, warnings and notes}
\usage{
missues(m)
}
\arguments{
\item{m}{An Rmonad}
}
\description{
Tabulates all errors, warnings and notes
}
\seealso{
Other from_Rmonad: \code{\link{esc}},
  \code{\link{mreport}}, \code{\link{mtabulate}}
}
