% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{mreport}
\alias{mreport}
\title{Convert a pipeline to Rmarkdown}
\usage{
mreport(m, section_prefix = "")
}
\arguments{
\item{m}{An Rmonad}
}
\description{
Convert a pipeline to Rmarkdown
}
\seealso{
Other monad-to-x: \code{\link{as_dgr_graph}},
  \code{\link{esc}}, \code{\link{missues}},
  \code{\link{mtabulate}}
}
