% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{mreport}
\alias{mreport}
\title{Convert a pipeline to Rmarkdown}
\usage{
mreport(m, section_prefix = "")
}
\arguments{
\item{m}{An Rmonad}

\item{section_prefix}{A prefix to add to all section headers (mostly for internal use)}
}
\description{
This function is experimental and may change completely in the future
}
\seealso{
Other from_Rmonad: \code{\link{as_dgr_graph}},
  \code{\link{esc}}, \code{\link{missues}},
  \code{\link{mtabulate}}
}
