% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{mreport}
\alias{mreport}
\title{Convert a pipeline to Rmarkdown}
\usage{
mreport(m)
}
\arguments{
\item{m}{An Rmonad}
}
\description{
STUB
}
\seealso{
Other from_Rmonad: \code{\link{esc}},
  \code{\link{missues}}, \code{\link{mtabulate}}
}
