% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_monad-to-x.R
\name{mreport}
\alias{mreport}
\title{Convert a pipeline to Rmarkdown}
\usage{
mreport(m)
}
\arguments{
\item{m}{An Rmonad}
}
\description{
This is currently a stub. It only pastes the docstrings and code blocks.
}
\seealso{
Other monad-to-x: \code{\link{as_dgr_graph}},
  \code{\link{esc}}, \code{\link{missues}},
  \code{\link{mtabulate}}
}
