% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monadic_eval.R
\name{mrun}
\alias{mrun}
\title{Run an expression, capture EWM, return Rmonad}
\usage{
mrun(expr, desc = NULL)
}
\arguments{
\item{expr}{An expression}

\item{desc}{A name to assign to the code slot}
}
\value{
Rmonad object
}
\description{
If the value is already an Rmonad, the existing value is returned.
}
