% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mrun}
\alias{mrun}
\title{Run an expression, capture EWM, return Rmonad}
\usage{
mrun(expr, desc = NULL)
}
\arguments{
\item{expr}{An expression}
}
\value{
Rmonad object
}
\description{
Run an expression, capture EWM, return Rmonad
}
