% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_monad-to-x.R
\name{mtabulate}
\alias{mtabulate}
\title{Make tabular summary of a monadic chain}
\usage{
mtabulate(m)
}
\arguments{
\item{m}{An Rmonad}
}
\description{
Make tabular summary of a monadic chain
}
\seealso{
Other monad-to-x: \code{\link{as_dgr_graph}},
  \code{\link{esc}}, \code{\link{missues}},
  \code{\link{mreport}}
}
