% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_monad-to-x.R
\name{mtabulate}
\alias{mtabulate}
\title{Make tabular summary of a pipeline}
\usage{
mtabulate(m, code = FALSE)
}
\arguments{
\item{m}{An Rmonad}

\item{code}{logical Should the code by included?}
}
\description{
Make tabular summary of a pipeline
}
\seealso{
Other from_Rmonad: \code{\link{esc}},
  \code{\link{missues}}, \code{\link{mreport}}
}
