% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cache.R
\name{noCache}
\alias{noCache}
\title{Represent a value that has been deleted}
\usage{
noCache()
}
\arguments{
\item{warn}{logical Emit a warning on access}
}
\description{
By default, the value of a node that has already been executed will be set
to this function.
}
