% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{note}
\alias{note}
\title{Append a note message onto the monad}
\usage{
note(m, s, force = FALSE)
}
\arguments{
\item{m}{A report monad}

\item{s}{A string describing a note}

\item{force}{logical, should we add the note even to a failed monad?}
}
\value{
A report monad with a new note appended
}
\description{
Append a note message onto the monad
}
