% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pass}
\alias{pass}
\title{Load a value into the report monad}
\usage{
pass(x)
}
\arguments{
\item{x}{The result of a successful computation}
}
\value{
The result wrapped in the report monad
}
\description{
Load a value into the report monad
}
\examples{
foo <- function(x) {
  if(x <= 0){
    fail(x, "x <= 0, cannot log")
  } else {
    pass(log(x))
  }
}
foo(-1)
foo(2)
}
