% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{Rmonad-class}
\alias{Rmonad-class}
\alias{Rmonad}
\title{The eponymous monad type}
\description{
The eponymous monad type
}
\section{Slots}{

\describe{
\item{\code{x}}{A list of 0 or 1 elements, containing either nothing or data}

\item{\code{OK}}{logical Is the monad passing?}

\item{\code{code}}{A string showing the function that created record's report}

\item{\code{error}}{An error in this this record}

\item{\code{warnings}}{Character vector of warnings}

\item{\code{notes}}{Character vector of notes}

\item{\code{doc}}{character vector documentation messages}

\item{\code{other}}{list of other things (currently includes time and space)}

\item{\code{parents}}{list of parent Rmonad objects}

\item{\code{.stored}}{logical (internal) whether an x value should be kept}
}}

