% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{rmonad_accessors}
\alias{rmonad_accessors}
\alias{has_code}
\alias{has_error}
\alias{has_doc}
\alias{has_warnings}
\alias{has_notes}
\alias{has_meta}
\alias{has_time}
\alias{has_mem}
\alias{has_value}
\alias{has_parents}
\alias{has_dependents}
\alias{has_prior}
\alias{has_nest}
\alias{has_summary}
\alias{get_parents}
\alias{get_dependents}
\alias{get_nest}
\alias{get_prior}
\alias{get_nest_depth}
\alias{get_value}
\alias{get_id}
\alias{get_OK}
\alias{get_code}
\alias{get_error}
\alias{get_warnings}
\alias{get_notes}
\alias{get_doc}
\alias{get_meta}
\alias{get_time}
\alias{get_mem}
\alias{get_summary}
\title{get, set, and append rmonad fields}
\usage{
has_code(m, ...)

has_error(m, ...)

has_doc(m, ...)

has_warnings(m, ...)

has_notes(m, ...)

has_meta(m, ...)

has_time(m, ...)

has_mem(m, ...)

has_value(m, ...)

has_parents(m, ...)

has_dependents(m, ...)

has_prior(m, ...)

has_nest(m, ...)

has_summary(m, ...)

get_parents(m, ...)

get_dependents(m, ...)

get_nest(m, ...)

get_prior(m, ...)

get_nest_depth(m, ...)

get_value(m, warn = TRUE, ...)

get_id(m, ...)

get_OK(m, ...)

get_code(m, ...)

get_error(m, ...)

get_warnings(m, ...)

get_notes(m, ...)

get_doc(m, ...)

get_meta(m, ...)

get_time(m, ...)

get_mem(m, ...)

get_summary(m, ...)
}
\arguments{
\item{m}{the rmonad}

\item{...}{Additional arguments}

\item{value}{value to replace or append current value}

\item{index}{The index of the node to get or set}
}
\description{
get, set, and append rmonad fields
}
