% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{rmonad_accessors}
\alias{rmonad_accessors}
\alias{has_code}
\alias{has_error}
\alias{has_doc}
\alias{has_warnings}
\alias{has_notes}
\alias{has_meta}
\alias{has_time}
\alias{has_mem}
\alias{has_value}
\alias{has_parents}
\alias{has_dependents}
\alias{has_prior}
\alias{has_nest}
\alias{has_summary}
\alias{ms_parents}
\alias{ms_dependents}
\alias{ms_nest}
\alias{ms_prior}
\alias{ms_nest_depth}
\alias{ms_value}
\alias{ms_id}
\alias{ms_OK}
\alias{ms_code}
\alias{ms_error}
\alias{ms_warnings}
\alias{ms_notes}
\alias{ms_doc}
\alias{ms_meta}
\alias{ms_time}
\alias{ms_mem}
\alias{ms_summary}
\title{get, set, and append rmonad fields}
\usage{
has_code(m, ...)

has_error(m, ...)

has_doc(m, ...)

has_warnings(m, ...)

has_notes(m, ...)

has_meta(m, ...)

has_time(m, ...)

has_mem(m, ...)

has_value(m, ...)

has_parents(m, ...)

has_dependents(m, ...)

has_prior(m, ...)

has_nest(m, ...)

has_summary(m, ...)

ms_parents(m, ...)

ms_dependents(m, ...)

ms_nest(m, ...)

ms_prior(m, ...)

ms_nest_depth(m, ...)

ms_value(m, warn = TRUE, ...)

ms_id(m, ...)

ms_OK(m, ...)

ms_code(m, ...)

ms_error(m, ...)

ms_warnings(m, ...)

ms_notes(m, ...)

ms_doc(m, ...)

ms_meta(m, ...)

ms_time(m, ...)

ms_mem(m, ...)

ms_summary(m, ...)
}
\arguments{
\item{m}{the rmonad}

\item{...}{Additional arguments}

\item{value}{value to replace or append current value}

\item{index}{The index of the node to get or set}
}
\description{
get, set, and append rmonad fields
}
