% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rmonad_accessors}
\alias{rmonad_accessors}
\alias{m_history}
\alias{m_OK}
\alias{m_value}
\alias{m_code}
\alias{m_error}
\alias{m_warnings}
\alias{m_notes}
\alias{m_doc}
\alias{m_branch}
\alias{m_history<-}
\alias{m_OK<-}
\alias{m_value<-}
\alias{m_code<-}
\alias{m_error<-}
\alias{m_warnings<-}
\alias{m_notes<-}
\alias{m_doc<-}
\alias{m_branch<-}
\alias{app_history}
\alias{app_warnings}
\alias{app_notes}
\alias{app_branch}
\title{get, set, and append rmonad fields}
\usage{
m_history(m)

m_OK(m)

m_value(m)

m_code(m)

m_error(m)

m_warnings(m)

m_notes(m)

m_doc(m)

m_branch(m)

m_history(m) <- value

m_OK(m) <- value

m_value(m) <- value

m_code(m) <- value

m_error(m) <- value

m_warnings(m) <- value

m_notes(m) <- value

m_doc(m) <- value

m_branch(m) <- value

app_history(m, value)

app_warnings(m, value)

app_notes(m, value)

app_branch(m, value)
}
\arguments{
\item{m}{the rmonad}

\item{value}{value to replace or append current value}
}
\description{
get, set, and append rmonad fields
}
