% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_accessors.R
\name{rmonad_accessors}
\alias{rmonad_accessors}
\alias{m_parents}
\alias{m_branch}
\alias{m_nest}
\alias{m_prior}
\alias{m_nest_depth}
\alias{m_value}
\alias{m_id}
\alias{m_OK}
\alias{m_code}
\alias{m_error}
\alias{m_warnings}
\alias{m_notes}
\alias{m_doc}
\alias{m_meta}
\alias{m_time}
\alias{m_mem}
\alias{m_summary}
\alias{m_OK<-}
\alias{m_value<-}
\alias{m_code<-}
\alias{m_error<-}
\alias{m_warnings<-}
\alias{m_notes<-}
\alias{m_doc<-}
\alias{m_meta<-}
\alias{m_time<-}
\alias{m_mem<-}
\alias{m_summary<-}
\alias{app_warnings}
\alias{app_notes}
\alias{m_parents<-}
\alias{m_nest<-}
\alias{m_nest_depth<-}
\alias{m_branch<-}
\alias{app_branch}
\alias{app_parents}
\title{get, set, and append rmonad fields}
\usage{
m_parents(m)

m_branch(m)

m_nest(m)

m_prior(m)

m_nest_depth(m)

m_value(m, ...)

m_id(m)

m_OK(m)

m_code(m)

m_error(m)

m_warnings(m)

m_notes(m)

m_doc(m)

m_meta(m)

m_time(m)

m_mem(m)

m_summary(m)

m_OK(m) <- value

m_value(m) <- value

m_code(m) <- value

m_error(m) <- value

m_warnings(m) <- value

m_notes(m) <- value

m_doc(m) <- value

m_meta(m) <- value

m_time(m) <- value

m_mem(m) <- value

m_summary(m) <- value

app_warnings(m, value)

app_notes(m, value)

m_parents(m) <- value

m_nest(m) <- value

m_nest_depth(m) <- value

m_branch(m) <- value

app_branch(m, value)

app_parents(m, value)
}
\arguments{
\item{m}{the rmonad}

\item{value}{value to replace or append current value}

\item{warn}{Warn if the accessed field does not exist (value was not cached)}
}
\description{
get, set, and append rmonad fields
}
