% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_magic.R
\name{splice_function}
\alias{splice_function}
\title{Take a monadic bind operation's result and splice histories}
\usage{
splice_function(f, m, ms)
}
\arguments{
\item{f}{The function}

\item{m}{The monadic result of running f(ms)}

\item{ms}{The list of inputs passed to f}
}
\description{
Take a monadic bind operation's result and splice histories
}
