% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{tag}
\alias{tag}
\title{Set the tag of an Rmonad object}
\usage{
tag(m, ..., index = m@head)
}
\arguments{
\item{m}{Rmonad object}

\item{...}{one or more tags for the given nodes}

\item{index}{character or integer vector, specifying the nodes which will be
assigned the new tag}
}
\value{
Rmonad object with new tags
}
\description{
Set the tag of an Rmonad object
}
\examples{
library(magrittr)
1 \%>>\% prod(2) \%>\% tag('a/b') \%>>\% prod(3) \%>\% get_tag

# to tag the first element in a pipeline, use `as_monad`:
as_monad(2, tag="x") \%>>\% sqrt \%>>\% get_tag

}
