% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{view}
\alias{view}
\title{Set the head of an Rmonad to a particular tag}
\usage{
view(m, tag)
}
\arguments{
\item{m}{Rmonad object}

\item{tag}{string specifying a single tag for one node in the pipeline}
}
\value{
Rmonad object with head reset
}
\description{
Set the head of an Rmonad to a particular tag
}
